move npmrc auth entirely to CI for npm publishing
Prior to this commit, the ai-toolkit's package-specific .npmrc file
would override the global ~/.npmrc file and cause authentication errors
in software that relied on the global npmrc auth (such as ai-toolkit itself,
as well as spec-workflow-mcp). Now, we delete the package-specific .npmrc file
so there is no conflict. The package-specific .npmrc file is created ephemerally
in every publish CI run, and that is th eonly place its needed